backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openpyxl==3.1.2
pywin32==307
python-dotenv==1.0.0
schedule==1.2.2
pandas>=2.0.0
rembg==2.0.62
onnxruntime>=1.16.0
psycopg2-binary==2.9.11
Flask-SQLAlchemy==3.0.5
APScheduler==3.10.4
fastapi==0.103.1
uvicorn==0.23.2
python-socketio==5.9.0
sqlalchemy>=2.0.0